add new code about calculadora JavaFx
