+++ src/org/neosuniversity/unidad3/calculadora/calculadora.fxml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextArea?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.BorderPane?>
+
+<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.121" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.neosuniversity.unidad3.calculadora.CalculadoraController">
+   <center>
+      <AnchorPane prefHeight="400.0" prefWidth="632.0" BorderPane.alignment="CENTER">
+         <children>
+            <TextField fx:id="txtOperadorA" layoutX="91.0" layoutY="31.0" promptText="Solo numeros" />
+            <TextField fx:id="txtOperadorB" layoutX="90.0" layoutY="81.0" promptText="Solo numeros" />
+            <TextField fx:id="txtResultado" editable="false" layoutX="222.0" layoutY="178.0" prefHeight="25.0" prefWidth="140.0" />
+            <Button layoutX="47.0" layoutY="121.0" mnemonicParsing="false" onAction="#actionSuma" prefHeight="25.0" prefWidth="39.0" text="+" />
+            <Button layoutX="106.0" layoutY="121.0" mnemonicParsing="false" onAction="#actionResta" prefHeight="25.0" prefWidth="41.0" text="-" />
+            <Button layoutX="164.0" layoutY="121.0" mnemonicParsing="false" onAction="#actionMultiplicacion" prefHeight="25.0" prefWidth="48.0" text="*" />
+            <Button layoutX="231.0" layoutY="121.0" mnemonicParsing="false" onAction="#actionDivision" prefHeight="25.0" prefWidth="49.0" text="/" />
+            <Label layoutX="17.0" layoutY="37.0" text="operador A" />
+            <Label layoutX="21.0" layoutY="87.0" text="operador B" />
+            <Label layoutX="21.0" layoutY="182.0" text="Resultado" />
+            <Button layoutX="137.0" layoutY="226.0" mnemonicParsing="false" onAction="#actionlimpiar" text="Limpiar" />
+            <TextArea fx:id="txtOperacion" editable="false" layoutX="86.0" layoutY="172.0" prefHeight="37.0" prefWidth="106.0" />
+            <Label layoutX="199.0" layoutY="182.0" text="=" />
+         </children>
+      </AnchorPane>
+   </center>
+</BorderPane>
